requirements.txt
streamlit
pandas
python-dateutil
sqlalchemy
psycopg2-binary
